chore: pages compoenents renaming
