+++ crowdin/messages.json
@@ -1973,15 +1973,15 @@
   "-1755817114": "Previously known vulnerable libraries without a working proof of concept (PoC)",
   "-709804513": "Output from automated vulnerability scanners without a PoC to demonstrate a specific vulnerability",
   "-1176025177": "Out of scope vulnerabilities",
-  "-1049724201": "Scope",
-  "-631514154": "Only Deriv assets are covered in this program. If you discover bugs on third-party apps, please report them to the respective owners. You may submit bug reports related to the following aspects of our business:",
-  "-908674373": "Please read and understand the Deriv Bug Bounty Program’s <0>terms and conditions</0> before you participate in the program.",
-  "-1425332024": "Bug bounty program",
   "-1372358243": "Critical",
   "-945348451": "Up to $5,000",
   "-1042333567": "Medium",
   "-1429914047": "Low",
   "-2147235454": "We’ll reward reports on low-level vulnerabilities <0>only</0> if they help us fix severe security issues, and we’ll decide the reward amount on a case-to-case basis. ",
+  "-1049724201": "Scope",
+  "-631514154": "Only Deriv assets are covered in this program. If you discover bugs on third-party apps, please report them to the respective owners. You may submit bug reports related to the following aspects of our business:",
+  "-908674373": "Please read and understand the Deriv Bug Bounty Program’s <0>terms and conditions</0> before you participate in the program.",
+  "-1425332024": "Bug bounty program",
   "-1131891395": "Important businesses",
   "-1541452017": "Our payment site: cashier.deriv.com",
   "-1441452782": "Our login site: oauth.deriv.com",
