+++ src/pages/bug-bounty/_lazy-load.ts
@@ -1,4 +1,4 @@
 import Loadable from '@loadable/component'
 
-export const Rewards = Loadable(() => import('./rewards'))
+export const Rewards = Loadable(() => import('./_rewards'))
 export const SubmitABugReport = Loadable(() => import('./_submit_a_bug_report'))
