+++ src/pages/bug-bounty/_rewards.tsx
