+++ src/pages/who-we-are/_OurOffices.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { graphql, useStaticQuery } from 'gatsby'
 import { desktop_pins, mobile_pins, our_offices_count } from './_data'
-import MapPin from './components/map-pin'
+import MapPin from './components/_map-pin'
 import { localize } from 'components/localization'
 import { SectionContainer, CssGrid, Desktop, Mobile, Flex } from 'components/containers'
 import { Header, Text, BackgroundImage } from 'components/elements'
