+++ src/pages/who-we-are/components/_leaders.tsx
@@ -2,7 +2,7 @@ import React, { MouseEventHandler, ReactElement, useState } from 'react'
 import styled from 'styled-components'
 import { graphql, useStaticQuery } from 'gatsby'
 import type { ImageDataLike } from 'gatsby-plugin-image'
-import Modal from './modal'
+import Modal from './_modal'
 import { Localize } from 'components/localization'
 import { Flex } from 'components/containers'
 import { QueryImage, ImageWrapper } from 'components/elements'
