+++ src/pages/who-we-are/components/_map-pin.tsx
