+++ src/pages/who-we-are/components/_modal.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from 'react'
 import styled from 'styled-components'
-import { ModalFlex } from './leaders'
+import { ModalFlex } from './_leaders'
 import { LocalizedLink } from 'components/localization'
 import { Header } from 'components/elements'
 import Linkedin from 'images/svg/who-we-are/linkedin.svg'
